feat[#10] : 보상 트랜잭션 실패 시 트랜잭션 롤백 로직 및 에러 로깅 추가

+++ app/services/log_reprocessor_service.py
@@ -11,20 +11,61 @@ async def retry_failed_logs(mongo_client: AsyncIOMotorClient):
     repo = ManagedActionLogRepository(mongo_client)
     weight_repo = UserWeightRepository(mongo_client)
     action_log_repo = ActionLogRepository(mongo_client)
-    failed_logs = await repo.find_failed_logs()
+
+    try: 
+        failed_logs = await repo.find_failed_logs()
+    except Exception as e:
+            logging.error(
+                f"[actionlog_status_update_failed] 보상트랜잭션 실패, log_id: {log['_id']}, error: {e}"
+            )
+            return
 
     for log in failed_logs:
-        try:
-            weight = await calc_weight(log)
-            await weight_repo.update_user_weights_from_log(log, weight)
-            await repo.delete_by_id(log["_id"])
-            await action_log_repo.update_status_to_success(log["_id"])
 
-            print(f"✅ 재처리 성공: {log['_id']}")
+        ## action 로그를 확인하여 이미 성공적으로 처리된 로그는 건너뛰기 (중복 가중치 업데이트 방지)
+        try:
+            status = await action_log_repo.get_status_by_id(log["_id"])
+
+            if status == "SUCCESS":
+                await repo.delete_by_id(log["_id"]) # 이미 성공적으로 처리된 로그는 삭제
+                print(f"❗ 이미 가중치가 업데이트된 로그입니다: {log['_id']}")
+                continue
         except Exception as e:
             logging.error(
                 f"[actionlog_status_update_failed] 보상트랜잭션 실패, log_id: {log['_id']}, error: {e}"
             )
+            continue
+
+        try:
+            weight = await calc_weight(log)
+            await weight_repo.update_user_weights_from_log(log, weight)
+        except Exception as e:
+            logging.error(
+                f"[actionlog_status_update_failed] 보상트랜잭션 실패, log_id: {log['_id']}, error: {e}"
+            )
+            continue
+
+        try:
+            await repo.delete_by_id(log["_id"])
+        except Exception as e:
+            logging.error(
+                f"[actionlog_status_update_failed] 보상트랜잭션 실패, log_id: {log['_id']}, error: {e}"
+            )
+            await weight_repo.decrease_user_weights_from_log(log, weight) 
+            continue
+
+        try:
+            await action_log_repo.update_status_to_success(log["_id"])
+        except Exception as e:
+            logging.error(
+                f"[actionlog_status_update_failed] 보상트랜잭션 실패, log_id: {log['_id']}, error: {e}"
+            )
+            await weight_repo.decrease_user_weights_from_log(log, weight) 
+            continue
+
+
+        print(f"✅ 재처리 성공: {log['_id']}")
+
 
 
 async def calc_weight(log: dict):
